Refactor[bmqtsk]: decouple AlarmLog observer and alarm macro
`bmqtsk_alarmlog.h` header contains both alarm log macro and `bmqtsk::AlarmLog` ball log
observer implementation. It is included in 40 cpp files, and most of these sources do not
need `bmqtsk::AlarmLog`. Ideally, they should not even know that this class exists.
This commit splits macro used for triggering alarms and alarm log observer class, also
renaming it to `bmqtsk::AlarmLogObserver`, similar to `bmqtsk::SysLogObserver`
and `bmqtsk::ConsoleObserver`.
