+++ src/groups/bmq/bmqtsk/bmqtsk_alarmlog.cpp
@@ -1,4 +1,4 @@
-// Copyright 2017-2023 Bloomberg Finance L.P.
+// Copyright 2026 Bloomberg Finance L.P.
 // SPDX-License-Identifier: Apache-2.0
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,97 +16,3 @@
 #include <bmqtsk_alarmlog.h>
 
 #include <bmqscm_version.h>
-
-#include <bmqu_memoutstream.h>
-#include <bmqu_stringutil.h>
-
-// BDE
-#include <ball_severity.h>
-#include <bdlt_timeunitratio.h>
-#include <bsl_iostream.h>
-#include <bsla_annotations.h>
-#include <bslma_allocator.h>
-#include <bslmt_lockguard.h>
-#include <bsls_performancehint.h>
-#include <bsls_timeutil.h>
-
-namespace BloombergLP {
-namespace bmqtsk {
-
-namespace {
-const int k_ALARM_INTERVAL = 60;
-// Minimum interval, in seconds, between two alarms from the same
-// source.
-
-/// Generate the alarm for the specified `category` with the specified
-/// `message` by writing to stderr.
-void generateAlarm(const bslstl::StringRef& category,
-                   const bslstl::StringRef& message)
-{
-    bsl::cerr << category << ": " << message << '\n' << bsl::flush;
-}
-
-}  // close unnamed namespace
-
-// --------------
-// class AlarmLog
-// --------------
-
-AlarmLog::AlarmLog(bslma::Allocator* allocator)
-: d_allocator_p(allocator)
-, d_lastAlarmTime(allocator)
-{
-    // NOTHING
-}
-
-AlarmLog::~AlarmLog()
-{
-    // NOTHING (required because of virtual class)
-}
-
-void AlarmLog::publish(const ball::Record&     record,
-                       BSLA_MAYBE_UNUSED const ball::Context& context)
-{
-    if (BSLS_PERFORMANCEHINT_PREDICT_LIKELY(
-            (record.fixedFields().severity() > ball::Severity::e_ERROR) ||
-            (record.customFields().length() == 0))) {
-        // We are only interested in ERROR severity, and only the ones which
-        // have a userField (the alarm type) set.
-        return;  // RETURN
-    }
-
-    BSLS_PERFORMANCEHINT_UNLIKELY_HINT;
-
-    bslmt::LockGuard<bslmt::Mutex> guard(&d_mutex);
-
-    const bsl::string& alarmType = record.customFields()[0].theString();
-
-    // Raw version used, no throttling
-    if (bmqu::StringUtil::startsWith(alarmType, "RAW_")) {
-        // We are stripping out the 'RAW_' prefix from the alarm identification
-        // string.
-        const bslstl::StringRef categoryStr(alarmType.c_str() + 4,
-                                            alarmType.length() - 4);
-        generateAlarm(categoryStr, record.fixedFields().messageRef());
-        return;  // RETURN
-    }
-
-    // Build the log source unique identifier (used for rate-limiting alarms)
-    // by concatenating the file and the line.
-    bmqu::MemOutStream ss(d_allocator_p);
-    ss << record.fixedFields().fileName() << ":"
-       << record.fixedFields().lineNumber() << bsl::ends;
-
-    bsls::Types::Int64& last = d_lastAlarmTime[ss.str()];
-    bsls::Types::Int64  now  = bsls::TimeUtil::getTimer();
-
-    if ((last == 0) ||
-        ((now - last) >=
-         k_ALARM_INTERVAL * bdlt::TimeUnitRatio::k_NANOSECONDS_PER_SECOND)) {
-        last = now;  // Update alarm last time
-        generateAlarm(alarmType, record.fixedFields().messageRef());
-    }
-}
-
-}  // close package namespace
-}  // close enterprise namespace
