+++ src/groups/bmq/bmqtsk/bmqtsk_alarmlogobserver.cpp
@@ -1,0 +1,112 @@
+// Copyright 2026 Bloomberg Finance L.P.
+// SPDX-License-Identifier: Apache-2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <bmqtsk_alarmlogobserver.h>
+
+#include <bmqscm_version.h>
+
+#include <bmqu_memoutstream.h>
+#include <bmqu_stringutil.h>
+
+// BDE
+#include <ball_severity.h>
+#include <bdlt_timeunitratio.h>
+#include <bsl_iostream.h>
+#include <bsla_annotations.h>
+#include <bslma_allocator.h>
+#include <bslmt_lockguard.h>
+#include <bsls_performancehint.h>
+#include <bsls_timeutil.h>
+
+namespace BloombergLP {
+namespace bmqtsk {
+
+namespace {
+const int k_ALARM_INTERVAL = 60;
+// Minimum interval, in seconds, between two alarms from the same
+// source.
+
+/// Generate the alarm for the specified `category` with the specified
+/// `message` by writing to stderr.
+void generateAlarm(const bslstl::StringRef& category,
+                   const bslstl::StringRef& message)
+{
+    bsl::cerr << category << ": " << message << '\n' << bsl::flush;
+}
+
+}  // close unnamed namespace
+
+// ----------------------
+// class AlarmLogObserver
+// ----------------------
+
+AlarmLogObserver::AlarmLogObserver(bslma::Allocator* allocator)
+: d_allocator_p(allocator)
+, d_lastAlarmTime(allocator)
+{
+    // NOTHING
+}
+
+AlarmLogObserver::~AlarmLogObserver()
+{
+    // NOTHING (required because of virtual class)
+}
+
+void AlarmLogObserver::publish(const ball::Record&     record,
+                               BSLA_MAYBE_UNUSED const ball::Context& context)
+{
+    if (BSLS_PERFORMANCEHINT_PREDICT_LIKELY(
+            (record.fixedFields().severity() > ball::Severity::e_ERROR) ||
+            (record.customFields().length() == 0))) {
+        // We are only interested in ERROR severity, and only the ones which
+        // have a userField (the alarm type) set.
+        return;  // RETURN
+    }
+
+    BSLS_PERFORMANCEHINT_UNLIKELY_HINT;
+
+    bslmt::LockGuard<bslmt::Mutex> guard(&d_mutex);
+
+    const bsl::string& alarmType = record.customFields()[0].theString();
+
+    // Raw version used, no throttling
+    if (bmqu::StringUtil::startsWith(alarmType, "RAW_")) {
+        // We are stripping out the 'RAW_' prefix from the alarm identification
+        // string.
+        const bslstl::StringRef categoryStr(alarmType.c_str() + 4,
+                                            alarmType.length() - 4);
+        generateAlarm(categoryStr, record.fixedFields().messageRef());
+        return;  // RETURN
+    }
+
+    // Build the log source unique identifier (used for rate-limiting alarms)
+    // by concatenating the file and the line.
+    bmqu::MemOutStream ss(d_allocator_p);
+    ss << record.fixedFields().fileName() << ":"
+       << record.fixedFields().lineNumber() << bsl::ends;
+
+    bsls::Types::Int64& last = d_lastAlarmTime[ss.str()];
+    bsls::Types::Int64  now  = bsls::TimeUtil::getTimer();
+
+    if ((last == 0) ||
+        ((now - last) >=
+         k_ALARM_INTERVAL * bdlt::TimeUnitRatio::k_NANOSECONDS_PER_SECOND)) {
+        last = now;  // Update alarm last time
+        generateAlarm(alarmType, record.fixedFields().messageRef());
+    }
+}
+
+}  // close package namespace
+}  // close enterprise namespace
